+++ src/core/models/Worker.ts
